# models/swagger.json
{
    "swagger": "2.0",
    "openapi": "3.0",
    "info": {
        "version": "1.0.0",
        "title": "Todo API",
        "description": "API endpoints for managing todos",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "servers": [
        {
            "url": "/",
            "description": "Local development server"
        }
    ],
    "tags": [
        {
            "name": "Todos",
            "description": "Todo management endpoints"
        }
    ],
    "paths": {
        "/delbyID/{id}": {
            "delete": {
                "tags": [
                    "Todos"
                ],
                "summary": "Delete a todo by ID",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "ID of the todo"
                    }
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/delSuccessResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Todo not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TODOErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TODOErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/todos": {
            "post": {
                "tags": [
                    "Todos"
                ],
                "summary": "Create a new todo",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "body",
                        "required": false,
                        "schema": {
                            "$ref": "#/components/schemas/Todo"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Todo"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/todosbyID/{id}": {
            "put": {
              "summary": "Update a todo by ID",
              "parameters": [
                {
                  "name": "id",
                  "in": "path",
                  "description": "ID of the todo to update",
                  "required": true,
                  "type": "string"
                },
                {
                  "name": "body",
                  "in": "body",
                  "description": "Updated todo data",
                  "required": true,
                  "schema": {
                    "$ref": "#/components/schemas/TodoRequestBody"
                  }
                }
              ],
              "responses": {
                "200": {
                  "description": "Successfully updated the todo",
                  "schema": {
                    "$ref": "#/components/schemas/Todo"
                  }
                },
                "404": {
                  "description": "Todo ID doesn't exist",
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                },
                "500": {
                  "description": "Internal server error",
                  "schema": {
                    "$ref": "#/components/schemas/TodoRequestBody"
                  }
                }
              }
            }},
        "/GettodosbyID/{id}": {
            "get": {
                "summary": "Get a todo by ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the todo to retrieve",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved the todo",
                        "schema": {
                            "$ref": "#/components/schemas/TodoRequestBody"
                        }
                    },
                    "404": {
                        "description": "Todo ID not found",
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            }
        },
        
"/todosAllData": {
    "get": {
      "summary": "Get all todos with pagination",
      "parameters": [
        {
          "name": "limit",
          "in": "query",
          "description": "Number of todos to retrieve per page",
          "required": false,
          "type": "integer"
        },
        {
          "name": "page",
          "in": "query",
          "description": "Page number of the todos to retrieve",
          "required": false,
          "type": "integer"
        }
      ],
      "responses": {
        "200": {
          "description": "Successfully retrieved the todos",
          "schema": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Todo"
            }
          }
        },
        "500": {
          "description": "Internal server error",
          "schema": {
            "$ref": "#/components/schemas/Error"
          }
        }
      }
    }
}
,


        "components": {
            "schemas": {
                "delSuccessResponse": {
                    "type": "object",
                    "properties": {
                        "message": {
                            "type": "string"
                        }
                    }
                },
                "ErrorResponse": {
                    "type": "object",
                    "properties": {
                        "error": {
                            "type": "string"
                        }
                    }
                },
                "Todo": {
                    "type": "object",
                    "properties": {
                        "title": {
                            "type": "string"
                        },
                        "description": {
                            "type": "string"
                        },
                        "completed": {
                            "type": "boolean"
                        },
                        "createdat": {
                            "type": "string",
                            "format": "date-time"
                        },
                        "updatedat": {
                            "type": "string",
                            "format": "date-time"
                        }
                    },
                    "required": [
                        "title",
                        "description"
                    ]
                },
                "TODOErrorResponse": {
                    "type": "object",
                    "properties": {
                        "title": {
                            "type": "string"
                        },
                        "description": {
                            "type": "string"
                        },
                        "completed": {
                            "type": "boolean"
                        },
                        "createdat": {
                            "type": "string",
                            "format": "date-time"
                        },
                        "updatedat": {
                            "type": "string",
                            "format": "date-time"
                        }
                    },
                    "required": [
                        "title",
                        "description"
                    ]
                },
                "UpdateTodo": {
                    "type": "object",
                    "properties": {
                        "title": {
                            "type": "string"
                        },
                        "description": {
                            "type": "string"
                        },
                        "completed": {
                            "type": "boolean"
                        }
                    },
                    "required": [
                        "title",
                        "description",
                        "completed"
                    ]
                },
                "TodoUP": {
                    "type": "object",
                    "properties": {
                        "_id": {
                            "type": "string"
                        },
                        "title": {
                            "type": "string"
                        },
                        "description": {
                            "type": "string"
                        },
                        "completed": {
                            "type": "boolean"
                        }
                    }
                }
                ,
                "Error": {
                    "type": "object",
                    "properties": {
                        "error": {
                            "type": "string"
                        }
                    }
                },
                "TodoRequestBody": {
                    "type": "object",
                    "properties": {
                        "title": {
                            "type": "string"
                        },
                        "description": {
                            "type": "string"
                        },
                        "completed": {
                            "type": "boolean"
                        }
                    }
                }
                
                
            }
        }
    }
}